Upgrade to NExt jS and a few bugfixes
