+++ Server/package.json
@@ -1,0 +1,5 @@
+{
+  "devDependencies": {
+    "@types/jest": "^29.5.1"
+  }
+}
