--- Webapp/jest.preset.js
@@ -1,9 +1,0 @@
-module.exports = {
-	transform: {
-		"^.+\\.(t|j)sx?$": "esbuild-jest"
-	},
-
-	transformIgnorePatterns: [
-		"[/\\\\]node_modules[/\\\\].+\\.(cjs|js|jsx|ts|tsx)$"
-	]
-}
