+++ Webapp/src/components/competition/PhaseSelector.tsx
@@ -201,7 +201,11 @@ const AddPhase = ({ refetch }: { refetch: () => Promise<any> }) => {
 				/>
 			</Grid>
 			<Grid item xs={12}>
-				<Button variant="contained" fullWidth onClick={submitNewPhase}>
+				<Button
+					variant="contained"
+					fullWidth
+					onClick={() => void submitNewPhase()}
+				>
 					Add Phase
 				</Button>
 			</Grid>
