--- Webapp/src/components/formSpecs/index.ts
@@ -1,2 +1,0 @@
-export { devSchema, extendedDevSchema } from "./dev";
-
