--- Webapp/src/components/formSpecs/jsonschemas/test copy.json
@@ -1,10 +1,0 @@
-{
-  "title": "Example JsonSchema Form",
-  "type": "object",
-  "required": ["title"],
-  "properties": {
-    "title": { "type": "string", "title": "Title", "default": "A new task" },
-    "done": { "type": "boolean", "title": "Done?", "default": false },
-      "numb": { "type":     "number", "title": "Done?", "default": false }
-  }
-}
