--- Webapp/src/components/formSpecs/jsonschemas/test.json
@@ -1,9 +1,0 @@
-{
-  "title": "Example JsonSchema Form",
-  "type": "object",
-  "required": ["title"],
-  "properties": {
-    "title": { "type": "string", "title": "Title", "default": "A new task" },
-    "done": { "type": "boolean", "title": "Done?", "default": false }
-  }
-}
