--- Webapp/src/components/formSpecs/typescript/test copy.d.ts
@@ -1,19 +1,0 @@
-/* tslint:disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export type Title = string;
-export type Done = boolean;
-export type Done1 = Done2 & Done3;
-export type Done2 = boolean;
-export type Done3 = number;
-
-export interface ExampleJsonSchemaForm {
-  title: Title;
-  done?: Done;
-  numb?: Done1;
-  [k: string]: unknown;
-}
