--- Webapp/src/components/formSpecs/typescript/test.d.ts
@@ -1,15 +1,0 @@
-/* tslint:disable */
-/**
- * This file was automatically generated by json-schema-to-typescript.
- * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
- * and run json-schema-to-typescript to regenerate this file.
- */
-
-export type Title = string;
-export type Done = boolean;
-
-export interface ExampleJsonSchemaForm {
-  title: Title;
-  done?: Done;
-  [k: string]: unknown;
-}
