+++ Webapp/src/components/roles/scribe/Bonus.tsx
