+++ Webapp/src/components/roles/scribe/tests/Move.spec.tsx
@@ -1,5 +1,5 @@
 import { fireEvent, screen } from "@testing-library/react"
-import { renderWithProviders } from "../../../../../testUtils"
+import { renderWithProviders } from "../../../../testUtils"
 import { AvailableMoveDirections, MovePropsType } from "../Interfaces"
 import { MoveCard } from "../MoveCard"
